LPD-74529 Separate unrelated tests in productMenu and portletInPage
